Do not force static linking
If ./configure is run without options, dontpanic tool is linked statically
to the dontpanic library. If ./configure is invoked with --disable-static,
libtool will add RPATH pointing to a build-time directory into the
dontpanic tool. The path is obviously wrong after installing the files
into a final location (make install).

It's probably a bug in libtool. This patch fixes/work around it by not
enforcing dontpanic tool static linking inside the Makefile.am when
--disable-static was requested.

Otherwise if both static and dynamic libraries are built, the tool
will be linked statically.

Signed-off-by: Petr Písař <ppisar@redhat.com>